Backend/Chatboard/requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
flask-cors==4.0.0 